Use a predictable resultAlias value for each possible Tomcat ThreadPool and GlobalRequestProcessor.
http-nio and ajp-nio Thread Pools and GlobalRequestProcessor MBean's names are always suffixed with the port number (for instance http-nio-8080). When the port number is randomly assigned (running in a PaaS like CloudFoundry), metrics will have different names every times tomcat is (re)started. This can be an issue in case you have a metric quota on the backend (when using hosted service like hostedgraphite.com). Even in the absence of a quota, it will simply fillup the backend disk partition really quickly. Also the metrics will be shortlive, and won't be easy to aggregate in the long run.

In order to fix that, we match each individual ObjectName with its prefix (commonly "http-nio-" and "ajp-nio-" in Tomcat) followed with a wildcard for the port number we don't know in advance. The resultAlias will not contain the port number at all, but transform to "http-nio" and "ajp-nio" respectively.

Note that when using a wildcard in a litteral attribute matcher, the value should be enclosed in double-quotes.